@@ -14,15 +14,14 @@ description: A eyetipzer Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cached_network_image: 2.5.1
+  cupertino_icons: 1.0.3
+  cached_network_image: 3.0.0
 
   flutter_ijkplayer: 0.3.5+1
 
@@ -35,11 +34,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  json_serializable: 4.0.2
-
-
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
+  json_serializable: 4.1.3
 
 # The following section is specific to Flutter.
 flutter:
