@@ -1,8 +1,8 @@
 class Consumption {
-  int collectionCount;
-  int shareCount;
-  int replyCount;
-  int realCollectionCount;
+  int? collectionCount;
+  int? shareCount;
+  int? replyCount;
+  int? realCollectionCount;
 
   Consumption.fromJson(Map<String, dynamic> json) {
     collectionCount = json['collectionCount'];
