@@ -41,29 +41,29 @@
 ///    "views": 1043
 ///    },
 class GankBean {
-  String id;
-  String author;
-  String category;
-  String content;
-  String createdAt;
-  String desc;
-  int index;
-  bool isOriginal;
-  String license;
-  int likeCount;
-  int likeCounts;
-  String markdown;
-  String originalAuthor;
-  String publishedAt;
-  int stars;
-  int status;
-  List<String> tags;
-  String title;
-  String type;
-  String url;
-  int views;
-  String updatedAt;
-  List<String> images;
+  String? id;
+  String? author;
+  String? category;
+  String? content;
+  String? createdAt;
+  String? desc;
+  int? index;
+  bool? isOriginal;
+  String? license;
+  int? likeCount;
+  int? likeCounts;
+  String? markdown;
+  String? originalAuthor;
+  String? publishedAt;
+  int? stars;
+  int? status;
+  List<String>? tags;
+  String? title;
+  String? type;
+  String? url;
+  int? views;
+  String? updatedAt;
+  List<String>? images;
 
   GankBean.fromJson(Map<String, dynamic> json) {
     id = json['_id'];
