@@ -8,13 +8,13 @@ import 'package:flutter_base/utils/string_utils.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
 
 class VideoByCategoryProvider extends BaseListViewModel with ChangeNotifier {
-  VideoRepository _videoResposity;
-  RefreshController refreshController;
-  ACategory category;
+  late VideoRepository _videoResposity;
+  RefreshController? refreshController;
+  ACategory? category;
   bool refreshFailed = false;
 
   int startPage = 0;
-  Trending last;
+  Trending? last;
 
   VideoByCategoryProvider({this.refreshController, this.category}) {
     //refresh();
@@ -22,44 +22,46 @@ class VideoByCategoryProvider extends BaseListViewModel with ChangeNotifier {
   }
 
   List<VideoItem> getVideos() {
-    return data;
+    return data.cast();
   }
 
   Future refresh() async {
-    print("refresh:${refreshController.footerStatus},$_videoResposity");
+    print("refresh:${refreshController?.footerStatus},$_videoResposity");
     setPage(startPage);
 
-    Trending trending =
+    Trending? trending =
         await _videoResposity.loadTrending(startPage, last, category: category);
     last = trending;
-    if (trending.itemList == null || trending.itemList.length < 1) {
-      refreshController.loadNoData();
+    if (trending == null ||
+        trending.itemList == null ||
+        trending.itemList!.length < 1) {
+      refreshController?.loadNoData();
     } else {
       setData(trending.itemList);
-      refreshController.refreshCompleted();
+      refreshController?.refreshCompleted();
     }
 
     notifyListeners();
   }
 
   @override
-  Future loadData({int pn}) {}
+  Future loadData({int? pn}) async {}
 
-  Future loadMore({int pn}) async {
+  Future loadMore({int? pn}) async {
     if (getCount() < 1) {
       return refresh();
     }
-    Trending trendingb = last;
+    Trending? trendingb = last;
     if (trendingb == null || StringUtils.isEmpty(trendingb.nextPageUrl)) {
-      refreshController.loadNoData();
+      refreshController?.loadNoData();
       return null;
     }
 
-    Trending trending =
+    Trending? trending =
         await _videoResposity.loadTrending(-1, last, category: category);
     if (trending != null &&
         trending.itemList != null &&
-        trending.itemList.length > 0) {
+        trending.itemList!.length > 0) {
       addData(trending.itemList);
 
       setPage(page + 1);
