@@ -3,9 +3,9 @@ import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 
 class CategoryItem extends StatelessWidget {
-  CategoryItem({this.bean, this.onPressed}) : super();
+  CategoryItem({required this.bean, this.onPressed}) : super();
   final ACategory bean;
-  final VoidCallback onPressed;
+  final VoidCallback? onPressed;
 
   void detail(String bean) {}
 
