@@ -4,16 +4,16 @@ import 'package:flutter/material.dart';
 import 'package:flutter_base/utils/string_utils.dart';
 
 class VideoListItem extends StatelessWidget {
-  VideoListItem({this.bean, this.onPressed}) : super();
+  VideoListItem({required this.bean, this.onPressed}) : super();
   final VideoItem bean;
-  final VoidCallback onPressed;
+  final VoidCallback? onPressed;
 
   void detail(String bean) {}
 
   @override
   Widget build(BuildContext context) {
     var titleContainer;
-    if (bean.data.dataType == "TextCard") {
+    if (bean.data!.dataType == "TextCard") {
       titleContainer = Container(
         height: 45,
         padding: const EdgeInsets.all(8.0),
@@ -24,7 +24,7 @@ class VideoListItem extends StatelessWidget {
           children: <Widget>[
             Expanded(
               child: Text(
-                  '${StringUtils.isEmpty(bean.data.text) ? "" : bean.data.text}',
+                  '${StringUtils.isEmpty(bean.data!.text) ? "" : bean.data!.text}',
                   style: TextStyle(fontSize: 20.0, color: Colors.cyanAccent)),
             ),
           ],
@@ -38,28 +38,28 @@ class VideoListItem extends StatelessWidget {
           crossAxisAlignment: CrossAxisAlignment.start,
           mainAxisAlignment: MainAxisAlignment.spaceAround,
           children: <Widget>[
-            Text('${bean.data.title}',
+            Text('${bean.data!.title}',
                 style: TextStyle(fontSize: 20.0, color: Colors.blue)),
             Expanded(
               child: Text(
-                  '${StringUtils.isEmpty(bean.data.description) ? "" : bean.data.description}',
+                  '${StringUtils.isEmpty(bean.data!.description) ? "" : bean.data!.description}',
                   style: TextStyle(fontSize: 14.0, color: Colors.cyanAccent)),
             ),
           ],
         ),
       );
     }
-    if (bean.data.dataType == "TextCard") {
+    if (bean.data!.dataType == "TextCard") {
       return Card(
         //color: Colors.black26,
         margin: const EdgeInsets.all(4.0),
-        child: new Material(
-          child: new Ink(
+        child: Material(
+          child: Ink(
             //INK可以实现装饰容器，实现矩形  设置背景色
             color: Colors.black,
-            child: new InkWell(
+            child: InkWell(
               onTap: () {},
-              child: new Container(
+              child: Container(
                 child: Stack(
                   children: <Widget>[
                     titleContainer,
@@ -74,12 +74,12 @@ class VideoListItem extends StatelessWidget {
       return Card(
         //color: Colors.black26,
         margin: const EdgeInsets.all(4.0),
-        child: new Material(
-          child: new Ink(
+        child: Material(
+          child: Ink(
             //INK可以实现装饰容器，实现矩形  设置背景色
             //color: Colors.black,
-            child: new InkWell(
-              child: new Container(
+            child: InkWell(
+              child: Container(
                 child: Stack(
                   children: <Widget>[
                     Container(
@@ -87,7 +87,7 @@ class VideoListItem extends StatelessWidget {
                         borderRadius: BorderRadius.circular(5.0),
                         child: Image(
                           image: CachedNetworkImageProvider(
-                              '${bean.data.cover.feed}'),
+                              '${bean.data!.cover!.feed}'),
                           fit: BoxFit.cover,
                         ),
                       ),
@@ -106,9 +106,9 @@ class VideoListItem extends StatelessWidget {
       //return Card(
       //  margin: const EdgeInsets.all(4.0),
       //  child: Material(
-      //    child: new Ink(
+      //    child: Ink(
       //      color: Colors.black,
-      //      child: new InkWell(
+      //      child: InkWell(
       //        child: Stack(
       //          children: <Widget>[
       //            Container(
@@ -133,12 +133,12 @@ class VideoListItem extends StatelessWidget {
       //);
       //return Column(
       //    children: <Widget>[
-      //      new Center(
-      //        child:new Material(
-      //          child: new InkWell(
+      //      Center(
+      //        child:Material(
+      //          child: InkWell(
       //            onTap: (){
       //            },
-      //            child: new Container(//不要在这里设置背景色，for则会遮挡水波纹效果,如果设置的话尽量设置Material下面的color来实现背景色
+      //            child: Container(//不要在这里设置背景色，for则会遮挡水波纹效果,如果设置的话尽量设置Material下面的color来实现背景色
       //              margin: EdgeInsets.all(0.0),
       //              width: 300.0,
       //              height: 100.0,
@@ -148,15 +148,15 @@ class VideoListItem extends StatelessWidget {
       //          color: Colors.yellow,
       //        ),
       //      ),
-      //      new Center(
-      //        child: new Material(
-      //          child: new Ink(//INK可以实现装饰容器，实现矩形  设置背景色
+      //      Center(
+      //        child: Material(
+      //          child: Ink(//INK可以实现装饰容器，实现矩形  设置背景色
       //            color: Colors.black,
-      //            child:new InkWell(
+      //            child:InkWell(
       //              onTap: (){
       //                onPressed();
       //              },
-      //              child: new Container(
+      //              child: Container(
       //                width: 300.0,
       //                height: 100.0,
       //              ),
@@ -164,27 +164,27 @@ class VideoListItem extends StatelessWidget {
       //          ),
       //        ),
       //      ),
-      //      new Center(
-      //        child: new Material(
-      //          child: new Ink(//用ink圆角矩形
+      //      Center(
+      //        child: Material(
+      //          child: Ink(//用ink圆角矩形
 //    //            color: Colors.red,
-      //            decoration: new BoxDecoration(
+      //            decoration: BoxDecoration(
       //              /**
       //               * assert(color == null || decoration == null,'Cannot provide both a color and
       //               * a decoration\n'The color argument is just a shorthand
-      //               * for "decoration: new BoxDecoration(color: color)".')
+      //               * for "decoration: BoxDecoration(color: color)".')
       //               * 在dart中使用assert 语句来中断正常的执行流程。
       //               * “不能同时”使用Ink的变量color属性以及decoration属性，“两个只能存在一个”。
       //               */
       //              color: Colors.purple,
-      //              borderRadius:new BorderRadius.all(new Radius.circular(20.0)),
+      //              borderRadius:BorderRadius.all(Radius.circular(20.0)),
       //            ),
-      //            child: new InkWell(
-      //              borderRadius:new BorderRadius.circular(20.0),//给水波纹也设置同样的圆角
+      //            child: InkWell(
+      //              borderRadius:BorderRadius.circular(20.0),//给水波纹也设置同样的圆角
       //              onTap: (){
       //                onPressed();
       //              },
-      //              child: new Container(
+      //              child: Container(
       //                width: 300.0,
       //                height: 100.0,
       //              ),
@@ -192,15 +192,15 @@ class VideoListItem extends StatelessWidget {
       //          ),
       //        ),
       //      ),
-      //      new Center(
-      //        child: new Material(
-      //          child:new Ink(
-      //            decoration: new BoxDecoration(
+      //      Center(
+      //        child: Material(
+      //          child:Ink(
+      //            decoration: BoxDecoration(
       //              color: Colors.red,
-      //              borderRadius: new BorderRadius.all(new Radius.circular(30.0)),
+      //              borderRadius: BorderRadius.all(Radius.circular(30.0)),
       //            ),
-      //            child:new InkResponse(
-      //              borderRadius: new BorderRadius.all(new Radius.circular(30.0)),
+      //            child:InkResponse(
+      //              borderRadius: BorderRadius.all(Radius.circular(30.0)),
 //    //              highlightColor: Colors.yellowAccent,//点击或者toch控件高亮时显示的控件在控件上层,水波纹下层
       //              highlightShape: BoxShape.rectangle,//点击或者toch控件高亮的shape形状
       //              //.InkResponse内部的radius这个需要注意的是，我们需要半径大于控件的宽，如果radius过小，显示的水波纹就是一个很小的圆，
@@ -209,7 +209,7 @@ class VideoListItem extends StatelessWidget {
       //              containedInkWell: true,//true表示要剪裁水波纹响应的界面   false不剪裁  如果控件是圆角不剪裁的话水波纹是矩形
       //              onTap: (){
       //              },
-      //              child: new Container(//1.不能在InkResponse的child容器内部设置装饰器颜色，否则会遮盖住水波纹颜色的，containedInkWell设置为false就能看到是否是遮盖了。
+      //              child: Container(//1.不能在InkResponse的child容器内部设置装饰器颜色，否则会遮盖住水波纹颜色的，containedInkWell设置为false就能看到是否是遮盖了。
       //                width: 300.0,
       //                height: 100.0,
       //              ),
@@ -217,15 +217,15 @@ class VideoListItem extends StatelessWidget {
       //          ),
       //        ),
       //      ),
-      //      new Center(
-      //        child: new Material(
-      //          child:new Ink(
-      //            decoration: new BoxDecoration(
+      //      Center(
+      //        child: Material(
+      //          child:Ink(
+      //            decoration: BoxDecoration(
       //              color: Colors.red,
-      //              borderRadius: new BorderRadius.all(new Radius.circular(30.0)),
+      //              borderRadius: BorderRadius.all(Radius.circular(30.0)),
       //            ),
-      //            child:new InkResponse(
-      //              borderRadius: new BorderRadius.all(new Radius.circular(30.0)),
+      //            child:InkResponse(
+      //              borderRadius: BorderRadius.all(Radius.circular(30.0)),
       //              highlightColor: Colors.yellowAccent,//点击或者toch控件高亮时显示的控件在控件上层,水波纹下层
       //              highlightShape: BoxShape.rectangle,//点击或者toch控件高亮的shape形状
       //              //.InkResponse内部的radius这个需要注意的是，我们需要半径大于控件的宽，如果radius过小，显示的水波纹就是一个很小的圆，
@@ -234,7 +234,7 @@ class VideoListItem extends StatelessWidget {
       //              containedInkWell: true,//true表示要剪裁水波纹响应的界面   false不剪裁  如果控件是圆角不剪裁的话水波纹是矩形
       //              onTap: (){
       //              },
-      //              child: new Container(//1.不能在InkResponse的child容器内部设置装饰器颜色，否则会遮盖住水波纹颜色的，containedInkWell设置为false就能看到是否是遮盖了。
+      //              child: Container(//1.不能在InkResponse的child容器内部设置装饰器颜色，否则会遮盖住水波纹颜色的，containedInkWell设置为false就能看到是否是遮盖了。
       //                width: 300.0,
       //                height: 100.0,
       //              ),
